pyproject.toml
[tool.poetry]
name = "chat-playground-gemini"
version = "0.1.0"
description = "A unified chat interface for Cerebras and Gemini AI models with system prompt management"
authors = ["Nicky <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.28.0"
cerebras-cloud-sdk = "*"
google-generativeai = "^0.3.0"
python-dotenv = "^1.0.0"
google-genai = "^1.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"] 